+++ nfttests/nft_templates.go
@@ -1,4 +1,4 @@
-package tests
+package nfttests
 
 import (
 	"fmt"
